refactor: Rename `DownloadPublicationButton` to `DownloadFileButton` for reusability
